headlines: move view code into converse-headlines-view.js
