+++ src/templates/headline_panel.html
@@ -2,5 +2,5 @@
 <div class="d-flex controlbox-padded">
     <span class="w-100 controlbox-heading controlbox-heading--headline">{{{o.heading_headline}}}</span>
 </div>
-<div class="list-container list-container--headline"></div>
+<div class="list-container list-container--headline hidden"></div>
 <!-- </div> -->
